+++ docs/nav/release-notes.md
@@ -3,6 +3,7 @@
 ### v1.0.19
 
 - [📦 PyPI - Build 1.0.19](https://github.com/linux-profile/linux-profile/releases/tag/v1.0.19)
+- [📌 ISSUE-170 - Update install script](https://github.com/linux-profile/linux-profile/issues/170)
 - [📌 ISSUE-166 - More package manager support to MacOS](https://github.com/linux-profile/linux-profile/issues/166)
 - [📌 ISSUE-165 - More package manager support for programming languages](https://github.com/linux-profile/linux-profile/issues/165)
 - [📌 ISSUE-163 - Support for more terminal types for running scripts](https://github.com/linux-profile/linux-profile/issues/163)